// templates/react-fastapi/src/lib/api/schema.d.ts
export interface components {
    schemas: {
        /** CsvReport */
        CsvReport: {
            /** Columns */
            columns: string[];
            /** Rows */
            rows: number;
        };
        /** HTTPValidationError */
        HTTPValidationError: {
            /** Detail */
            detail?: components["schemas"]["ValidationError"][];
        };
        /** Health */
        Health: {
            /** Status */
            status: string;
        };
        /** Identity */
        Identity: {
            /** Subject */
            subject: string;
        };
        /** NoteIn */
        NoteIn: {
            /** Body */
            body: string;
        };
        /** NoteOut */
        NoteOut: {
            /** Body */
            body: string;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /** Id */
            id: number;
        };
        /** ValidationError */
        ValidationError: {
            /** Context */
            ctx?: Record<string, never>;
            /** Input */
            input?: unknown;
            /** Location */
            loc: (string | number)[];
            /** Message */
            msg: string;
            /** Error Type */
            type: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}